# UsersManager/src/main/java/com/ucx/usersmanager/manageusers/models/UserModel.java
/*
 * Copyright (C) 2022. Smart Bits GmbH. All rights reserved.
 * Created by Arbër Pllana on 03/03/2022.
 */

package com.ucx.usersmanager.manageusers.models;

import com.ucx.usersmanager.commons.models.BoBaseModel;

import javax.persistence.Entity;

@Entity(name = "StoreUsers")
public class UserModel extends BoBaseModel {
  private String firstName;
  private String lastName;
  private String email;
  private String password;

  public String getFirstName() {
    return firstName;
  }

  public void setFirstName(String firstName) {
    this.firstName = firstName;
  }

  public String getLastName() {
    return lastName;
  }

  public void setLastName(String lastName) {
    this.lastName = lastName;
  }

  public String getEmail() {
    return email;
  }

  public void setEmail(String email) {
    this.email = email;
  }

  public String getPassword() {
    return password;
  }

  public void setPassword(String password) {
    this.password = password;
  }
}